Blender API Test completed

See merge request !11

+++ src/rendering/BlenderAPI/BlenderObjects.py
@@ -3,6 +3,8 @@ import math
 import mathutils as mathU
 import itertools
 
+from rendering.BlenderAPI.BlenderExceptions import *
+
 def rotate(vector, quaternion):
     """
     utility function to rotate a vector, given a rotation in the form of a quaternion
@@ -28,7 +30,7 @@ def to_quaternion(w, x, y, z):
     m = math.sqrt(x ** 2 + y ** 2 + z ** 2)
     w = math.pi * w / 180.0
     if m == 0:
-        q = [0, 0, 0, 0]
+        q = mathU.Quaternion([0, 0, 0], 0)
     else:
         q = mathU.Quaternion([x / m, y / m, z / m], w)
     return q
@@ -50,18 +52,28 @@ class BlenderObject(object):
     """
 
     def __init__(self, location=(0, 0, 0), orientation=(0, 0, 0, 0), scale=(1, 1, 1), reference=None, **kwargs):
+
         if reference is None:
+
             bpy.ops.object.select_all(action='DESELECT')  # deselect everything
-            self.blender_create_operation(location, **kwargs)
+            self.blender_create_operation(**kwargs)
             assert len(bpy.context.selected_objects) == 1, "more than one selected objects!"
             # make sure the only selected object is the recently created object
             self.reference = bpy.context.selected_objects[0]
+
         else:
+
             self.reference = reference
+
+        valid = check_is_iter(location, 3) and check_is_iter(orientation, 4)
+        if not valid:
+            raise InvalidInputError("location input invalid")
+
+        self.set_location(*location)
         self.set_rot(*orientation)
         self.set_scale(scale)
 
-    def blender_create_operation(self, location):
+    def blender_create_operation(self):
         # Attention: Pure virtual, for subclass to implement
         raise NotImplementedError
 
@@ -76,6 +88,9 @@ class BlenderObject(object):
         set the scale of current object
         :param scale: 3-tuple specifying scale of the x,y,z axes
         """
+        valid = check_is_iter(scale, 3) and check_vector_non_negative(scale)
+        if not valid:
+            raise InvalidInputError("scale input invalid")
         self.reference.scale = scale
 
     def set_rot(self, w, x, y, z):
@@ -105,16 +120,25 @@ class BlenderObject(object):
         self.reference.rotation_mode = 'QUATERNION'
         q = to_quaternion(w, x, y, z)
         q = q * self.reference.rotation_quaternion
-        print(q)
         self.reference.rotation_quaternion = q
 
     def delete(self):
         """
         delete current object, by deleting its reference
         """
+        if self.reference is None:
+            return # object reference already deleted
         # deselect all
         bpy.ops.object.select_all(action='DESELECT')
         # selection
         self.reference.select = True
         # remove it
         bpy.ops.object.delete()
+        self.reference = None
+
+class BlenderTestObject(BlenderObject):
+    def __init__(self, location=(0, 0, 0), orientation=(0, 0, 0, 0), scale=(1, 1, 1), reference=None, **kwargs):
+        super(BlenderTestObject, self).__init__(location, orientation, scale, reference)
+
+    def blender_create_operation(self):
+        bpy.ops.object.add()
